Day 14: Extended Polymerization

+++ 2021/day1.cpp
@@ -7,7 +7,7 @@ std::pair<std::uintmax_t,std::uintmax_t> solve() {
   std::vector<int> w;
   int cur;
   while (std::cin >> cur) w.push_back(cur);
-  for (int i = 1; i < w.size(); i++) {
+  for (int i = 1; i < (int)w.size(); i++) {
     if (w[i-1] < w[i]) res.first++;
     if (w[i-3] < w[i] && i > 2) res.second++;
   }
